+++ app/domain/user.model.ts
@@ -3,20 +3,22 @@ import { uiLocales } from '@/i18n/locales';
 import { DEFAULT_LOCALE } from './locale.model';
 import { roleSchema } from './role.model';
 
-export const userSchema = z.object({
-  id: z.uuid(),
-  name: z.string().min(1).max(255),
-  email: z.string().email().max(255),
-  emailVerified: z.boolean().default(false),
-  image: z.string().nullable().default(null),
-  locale: z.string().nullable().default(DEFAULT_LOCALE),
-  role: roleSchema.nullable().default(null),
-  banned: z.boolean().nullable().default(false),
-  banReason: z.string().nullable().default(null),
-  banExpires: z.date().nullable().default(null),
-  createdAt: z.date(),
-  updatedAt: z.date(),
-});
+export const userSchema = z
+  .object({
+    id: z.string().min(1),
+    name: z.string().min(1).max(255),
+    email: z.string().email().max(255),
+    emailVerified: z.boolean().default(false),
+    image: z.string().nullable().default(null),
+    locale: z.string().nullable().default(DEFAULT_LOCALE),
+    role: roleSchema.nullable().default(null),
+    banned: z.boolean().nullable().default(false),
+    banReason: z.string().nullable().default(null),
+    banExpires: z.coerce.date().nullable().default(null),
+    createdAt: z.coerce.date(),
+    updatedAt: z.coerce.date(),
+  })
+  .strict();
 
 export type User = z.infer<typeof userSchema>;